fix(nuxt): get asyncData from context
BREAKING CHANGE: loaderContentFetch requires suspense
