+++ package.json
@@ -65,7 +65,7 @@
     "postcss": "^8.5.6",
     "postcss-cli": "^11.0.1",
     "postcss-html": "^1.8.0",
-    "prettier": "^3.7.4",
+    "prettier": "^3.9.3",
     "rimraf": "^6.1.2",
     "semantic-release": "^17.1.1",
     "semantic-release-monorepo": "^7.0.5",
@@ -102,5 +102,6 @@
   },
   "browserslist": [
     "defaults"
-  ]
+  ],
+  "dependencies": {}
 }
