+++ packages/components-vue/package.json
@@ -52,12 +52,6 @@
       "import": "./dist/plugin.js",
       "require": "./dist/plugin.cjs",
       "default": "./dist/plugin.js"
-    },
-    "./theme": {
-      "types": "./dist/composables/theme.d.ts",
-      "import": "./dist/theme.js",
-      "require": "./dist/theme.cjs",
-      "default": "./dist/theme.js"
     }
   },
   "dependencies": {
