+++ packages/components-vue/src/types/plugin.ts
@@ -3,12 +3,8 @@ import type { Extension } from "@codemirror/state";
 
 import type { iPluginOptions } from "@open-xamu-co/ui-common-types";
 
-import type { useAsyncDataFn } from "../composables/async";
-
 export type vComponent<P extends Record<string, any> = Record<string, any>> =
-	| VueComponent<P>
-	| FunctionalComponent<P>
-	| DefineComponent<P>;
+	VueComponent<P> | FunctionalComponent<P> | DefineComponent<P>;
 
 export interface iVuePluginOptions extends iPluginOptions<vComponent> {
 	/**
@@ -27,10 +23,6 @@ export interface iVuePluginOptions extends iPluginOptions<vComponent> {
 		 * Client only component
 		 */
 		clientOnly?: vComponent;
-		/**
-		 * Nuxt Async data
-		 */
-		useAsyncData?: typeof useAsyncDataFn;
 		/**
 		 * Nitro fetch
 		 */
