+++ packages/components-vue/src/types/props/index.ts
@@ -2,3 +2,4 @@ export * from "./base";
 export * from "./modal";
 export * from "./table";
 export * from "./value";
+export * from "./async";
