--- packages/components-vue/vite.config.d.ts
@@ -1,3 +1,0 @@
-declare const _default: import("vite").UserConfig;
-
-export default _default;
