+++ packages/components-vue/vite.config.ts
@@ -14,7 +14,7 @@ export default defineConfig({
 				// plugin
 				plugin: path.resolve(__dirname, "./src/plugin"),
 				// theme
-				theme: path.resolve(__dirname, "./src/composables/theme"),
+				// theme: path.resolve(__dirname, "./src/composables/theme"),
 			},
 			name: "@open-xamu-co/ui-components-vue",
 			formats: ["es", "cjs"],
