--- packages/components-vue/vitest.config.d.ts
@@ -1,3 +1,0 @@
-declare const _default: Record<string, any>;
-
-export default _default;
