--- packages/components-vue/vitest.shims.d.ts
@@ -1,1 +1,0 @@
-/// <reference types="@vitest/browser-playwright" />
