+++ packages/styles/src/base/_initialize.scss
@@ -94,7 +94,7 @@
 		h4,
 		h5,
 		h6,
-		th,
+		th span,
 		button,
 		strong {
 			// @defaults txtWeight
